Put slow mark in separate plugin
